Diamond Price Predictor using Machine Learning
The diamond dataset was obtained from the Kaggle website. The dataset used in this project can be downloaded from this website: https://www.kaggle.com/datasets/nancyalaswad90/diamonds-prices. This dataset contains the prices and other attributes of almost 54,000 diamonds. The diamond price was predicted using the best model. Initially, linear regression, decision trees, and random forests were used. I chose the Random Forest model as the best model among the three based on its MAE and MSLE scores on the test set. Then this random forest model is also able to predict the price of every diamond in the test set with an error of ±1.20% of the real price.